final-pjt-back/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cron-descriptor==1.4.5
crypto==1.4.1
cryptography==45.0.2
cycler==0.12.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.8.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
filelock==3.18.0
finnhub-python==2.4.23
fonttools==4.57.0
fsspec==2025.3.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.7
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib
mpmath==1.3.0
mysqlclient==2.1.1
Naked==0.1.32
networkx==3.2.1
numpy
oauthlib==3.2.2
openai==1.70.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy
seaborn==0.13.2
selenium==4.30.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.122
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zipp==3.22.0
